+++ mla/__init__.py
@@ -2,7 +2,9 @@
 # flake8: noqa
 from . import analysis
 from . import models
+from . import ThreeML_models
 from . import sources
 from . import spectral
 from . import test_statistics
+from . import ThreeML_test_statistics
 from . import time_profiles
