+++ mla/test_statistics.py
@@ -251,52 +251,3 @@ def i3_test_statistic(params: np.ndarray,
     if 'ns' in temp_params.dtype.names:
         ns_ratio = temp_params['ns'] / prepro.n_events
     return _i3_ts(sob, prepro, return_ns, ns_ratio)
-
-
-@dataclasses.dataclass
-class ThreeMLPreprocessing(_TdPreprocessing):
-    """Docstring"""
-    event_model: models.ThreeMLEventModel
-
-
-@dataclasses.dataclass
-class ThreeMLPreprocessor(_TdPreprocessor):
-    """Docstring"""
-    factory_type: ClassVar = ThreeMLPreprocessing
-
-
-def threeml_ps_test_statistic(params: np.ndarray,
-                              prepro: ThreeMLPreprocessing,
-                              return_ns: bool = False) -> float:
-
-    """(ThreeML version) Evaluates the ts for the given events and parameters
-
-    Calculates the test-statistic using a given event model, n_signal, and
-    gamma. This function does not attempt to fit n_signal or gamma.
-
-    Args:
-        params:
-        event_model:
-        events:
-        prepro:
-        return_ns:
-
-    Returns:
-        The overall test-statistic value for the given events and
-        parameters.
-    """
-    temp_params = rf.unstructured_to_structured(
-        params, dtype=prepro.params.dtype, copy=True)
-
-    sob_energy = prepro.event_model.get_energy_sob(
-        prepro.events[prepro['drop_index']])
-
-    sob = prepro.sob_spatial * _sob_time(params, prepro) * sob_energy
-
-    sob = prepro.sob_spatial * _sob_time(
-        temp_params, prepro) * sob_energy
-
-    ns_ratio = None
-    if 'ns' in temp_params.dtype.names:
-        ns_ratio = temp_params['ns'] / prepro.n_events
-    return _i3_ts(sob, prepro, return_ns, ns_ratio)
